+++ ims/ims.py
@@ -5,7 +5,7 @@ import sys
 
 import requests
 from bs4 import BeautifulSoup
-from piratebayCrawler import main
+from urlparse import urljoin
 DEBUG = False
 
 
@@ -50,6 +50,57 @@ def get_torrent_url(search_url):
     return torrent_url
 
 
+base_url = 'https://piratebays.co.uk/'
+
+def match(text, words):
+    ctr = 0
+    text = text.lower()
+    for word in words:
+        if word in text:
+            ctr += 1
+    if ctr >= len(words) - 1:
+        return True
+    else:
+        return False
+
+
+def get_page_link(name):
+    try:
+        url = base_url + 's/'
+        data = {'q': name, 'video': 'on', 'page': 0, 'orderby': 99}
+        r = requests.get(url, params = data)
+        soup = BeautifulSoup(r.text, 'html.parser')
+
+        searchBody = soup.find(id = 'searchResult')
+        rows = searchBody.find_all('tr')[1 : ]
+        if len(rows) > 0:
+            words = name.split(' ')
+            for r in rows:
+                div = r.find_all('td')[1]
+                a_tags = div.find_all('a')
+                title = a_tags[0].text
+                if match(title, words):
+                    link = a_tags[1].get('href')
+                    page_link = urljoin(base_url, link)
+                    return page_link
+    except Exception as e:
+        print e
+        return None
+
+
+def get_magnet_link(link):
+    try:
+        r = requests.get(link)
+        soup = BeautifulSoup(r.text, 'html.parser')
+        div = soup.find(class_ = 'download')
+        a_tag = div.find('a')
+        magnet_link = a_tag.get('href')
+        return magnet_link
+    except Exception as e:
+        print e
+        return None
+
+
 test_system()
 movie = get_input()
 url = 'http://dx-torrente.com/usearch/' + movie
@@ -62,13 +113,14 @@ try:
 except Exception as e:
     #print e
     print 'Taking a bit longer....Please wait..'
-    if sys.argv[0] == 'ims.py':
-        movie_name = ' '.join(sys.argv[2 : ])
-    else:
-        movie_name = ' '.join(sys.argv[1 : ])
-    print movie_name
-    main(movie_name)
-    exit()
+    movie_name = ' '.join(sys.argv[2 : ])
+    name = movie_name.lower()
+    link = get_page_link(name)
+    if link:
+        torrent_url = get_magnet_link(link)
+    if torrent_url == '':
+        exit()
+
 if torrent_url:
     print ('Streaming Torrent: ' + torrent_url)
     os.system('peerflix ' + torrent_url + ' -a --vlc')
